Update OreFinder LUA Code
improved search algorithm.  It now waits for 3 measurements before providing a marker, also it checks if you are giving a measurement from a location more than 25% beyond the range of the first measurement, this can indicate that you are tracking a different node than the first one.  If this condition is detected the out of range measurement is ignored so you have a chance to take a better one.